+++ src/main/java/nextstep/member/application/JwtTokenProvider.java
@@ -10,6 +10,7 @@ import java.util.Date;
 public class JwtTokenProvider {
     @Value("${security.jwt.token.secret-key}")
     private String secretKey;
+
     @Value("${security.jwt.token.expire-length}")
     private long validityInMilliseconds;
