+++ src/main/java/nextstep/member/application/dto/TokenResponse.java
@@ -1,16 +1,18 @@
 package nextstep.member.application.dto;
 
 public class TokenResponse {
-    private String accessToken;
+    private final String accessToken;
 
-    public TokenResponse() {
+    private TokenResponse(String accessToken) {
+        this.accessToken = accessToken;
     }
 
-    public TokenResponse(String accessToken) {
-        this.accessToken = accessToken;
+    public static TokenResponse of(String accessToken) {
+        return new TokenResponse(accessToken);
     }
 
     public String getAccessToken() {
         return accessToken;
     }
 }
+
