+++ src/main/java/nextstep/member/domain/AuthService.java
@@ -1,0 +1,9 @@
+package nextstep.member.domain;
+
+import nextstep.member.application.dto.TokenRequest;
+import nextstep.member.application.dto.TokenResponse;
+
+public interface AuthService {
+    TokenResponse login(TokenRequest request);
+}
+
