--- src/main/java/nextstep/member/domain/TokenService.java
@@ -1,9 +1,0 @@
-package nextstep.member.domain;
-
-import nextstep.member.application.dto.TokenResponse;
-
-public interface TokenService {
-    TokenResponse createToken(String email, String password);
-
-    TokenResponse createTokenFromGithub(String code);
-}
