+++ src/main/java/nextstep/member/domain/UserDetails.java
@@ -1,0 +1,7 @@
+package nextstep.member.domain;
+
+public interface UserDetails {
+    String getPassword();
+
+    String getUsername();
+}
