+++ src/main/java/nextstep/member/domain/UserDetailsService.java
@@ -1,0 +1,7 @@
+package nextstep.member.domain;
+
+import nextstep.member.UsernameNotFoundException;
+
+public interface UserDetailsService {
+    UserDetails loadUserByUsername(String username) throws UsernameNotFoundException;
+}
