--- src/main/java/nextstep/member/tobe/AuthConfig.java
@@ -1,23 +1,0 @@
-package nextstep.member.tobe;
-
-import java.util.List;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import nextstep.member.application.JwtTokenProvider;
-import nextstep.member.ui.AuthenticationPrincipalArgumentResolver;
-
-@Configuration
-public class AuthConfig implements WebMvcConfigurer {
-    private JwtTokenProvider jwtTokenProvider;
-
-    public AuthConfig(JwtTokenProvider jwtTokenProvider) {
-        this.jwtTokenProvider = jwtTokenProvider;
-    }
-
-    @Override
-    public void addArgumentResolvers(List argumentResolvers) {
-        argumentResolvers.add(new AuthenticationPrincipalArgumentResolver(jwtTokenProvider));
-    }
-}
