--- src/main/java/nextstep/member/tobe/AuthenticationException.java
@@ -1,8 +1,0 @@
-package nextstep.member.tobe;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.UNAUTHORIZED)
-public class AuthenticationException extends RuntimeException {
-}
