+++ src/main/resources/application.properties
@@ -1,8 +1,13 @@
 spring.jpa.properties.hibernate.show_sql=true
 spring.jpa.properties.hibernate.format_sql=true
 
+security.jwt.token.secret-key=atdd-secret-key
+security.jwt.token.expire-length=3600000
+
+github.client-id=client-id
+github.client-secret=client-secret
+github.token-url=https://github.com/login/oauth/access_token
+github.user-url=https://api.github.com
+
 logging.level.org.hibernate.SQL=DEBUG
 logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
-
-security.jwt.token.secret-key= atdd-secret-key
-security.jwt.token.expire-length= 3600000
